Split UserModal into CreateUserModal and EditUserModal
- Updates the relevant unit tests
- Simplifies the components to make them easier to add to and improve
